TASK: Get rid of fusion implementations for Form and Field and use `@context` plus `@ignoreProperties` instead
