+++ Classes/Eel/FormHelper.php
@@ -45,6 +45,77 @@ class FormHelper implements ProtectedContextAwareInterface
     protected $hashService;
 
     /**
+     * Create a form definition object
+     *
+     * @param ActionRequest|null $request
+     * @param string|null $fieldNamePrefix
+     * @param mixed|null $data
+     * @return Form
+     */
+    public function createForm(ActionRequest $request = null, string $fieldNamePrefix = null, $data = null): Form
+    {
+        return new Form(
+            $request,
+            $data,
+            $fieldNamePrefix ?: ($request ? $request->getArgumentNamespace() : ''),
+            $request ? $request->getInternalArgument('__submittedArguments') : [],
+            $request ? $request->getInternalArgument('__submittedArgumentValidationResults') : new Result()
+        );
+    }
+
+    /**
+     * Create a field definition object
+     *
+     * @param Form|null $form
+     * @param string $name
+     * @param bool $multiple
+     * @return Field
+     */
+    public function createField(Form $form = null, string $name = null, bool $multiple = false): Field
+    {
+        if (!$name) {
+            return new Field(null, null, null);
+        }
+
+        // render fieldName
+        if ($form && $form->getFieldNamePrefix()) {
+            $fieldName = $this->prefixFieldName($name, $form->getFieldNamePrefix());
+        } else {
+            $fieldName = $name;
+        }
+        if ($multiple) {
+            $fieldName .= '[]';
+        }
+
+        // create property path from fieldname
+        $path = $this->fieldNameToPath($name);
+
+        // determine value, according to the following algorithm:
+        if ($form && $form->getResult() !== null && $form->getResult()->hasErrors()) {
+            // 1) if a validation error has occurred, pull the value from the submitted form values.
+            $fieldValue = ObjectAccess::getPropertyPath($form->getSubmittedValues(), $path);
+        } elseif ($path && $form && $form->getData()) {
+            // 2) else, if "property" is specified, take the value from the bound object.
+            $fieldValue = ObjectAccess::getPropertyPath($form->getData(), $path);
+        } else {
+            $fieldValue = null;
+        }
+
+        // determine ValidationResult for the single property
+        $fieldResult = null;
+        if ($form && $form->getResult() && $form->getResult()->hasErrors()) {
+            $fieldResult = $form->getResult()->forProperty($path);
+        }
+
+        return new Field(
+            $fieldName,
+            $fieldValue,
+            $multiple,
+            $fieldResult
+        );
+    }
+
+    /**
      * Calculate the hidden fields for the given form content as key-value array
      *
      * @param ActionRequest $request
